реализован таск для заказчика и публичный

+++ interal/customer/model.go
@@ -6,9 +6,9 @@ import (
 )
 
 type Customer struct {
-	Money  int                `json:"money,omitempty" json:"money,omitempty"`
-	Loader []loader.Loader    `json:"loader,omitempty" json:"loader,omitempty"`
-	Tasks  map[string]float64 `json:"tasks,omitempty" json:"tasks,omitempty"`
+	Money  int             `json:"money,omitempty" json:"money,omitempty"`
+	Loader []loader.Loader `json:"loader,omitempty" json:"loader,omitempty"`
+	Tasks  map[string]int  `json:"tasks,omitempty" json:"tasks,omitempty"`
 }
 
 func GenerateCustomer() *Customer {
